audit_site_1/requirements.txt
appscript==1.2.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-debug-toolbar==3.2.4
docx2pdf==0.1.8
docxcompose==1.3.4
docxtpl==0.15.2
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
numpy==1.22.1
pandas==1.4.0
Pillow==9.0.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
